refactor: extract abuse report handlers to layered architecture
- Create api/reports/create.rs
- Create api/reports/admin.rs
- Create repositories/report_repository.rs
- Remove 6 SQL functions from db/queries.rs
- Remove 5 handlers from router.rs
