+++ src/api/mod.rs
@@ -4,6 +4,7 @@ pub mod billing;
 pub mod keys;
 pub mod links;
 pub mod orgs;
+pub mod reports;
 pub mod settings;
 pub mod tags;
 pub mod title_fetch;
