+++ src/repositories/mod.rs
@@ -17,12 +17,14 @@
 pub mod analytics_repository;
 pub mod billing_repository;
 pub mod product_repository;
+pub mod report_repository;
 pub mod settings_repository;
 pub mod tag_repository;
 pub mod user_repository;
 pub use analytics_repository::AnalyticsRepository;
 pub use billing_repository::BillingRepository;
 pub use product_repository::ProductRepository;
+pub use report_repository::ReportRepository;
 pub use settings_repository::SettingsRepository;
 pub use tag_repository::TagRepository;
 pub use user_repository::UserRepository;
