+++ src/router.rs
@@ -3101,424 +3101,6 @@ pub async fn handle_admin_update_user_role(
 }
 
 #[utoipa::path(
-    post,
-    path = "/api/reports/links",
-    tag = "Reports",
-    summary = "Report a link for abuse",
-    description = "Submits an abuse report for a link. Accepts both authenticated and anonymous submissions. Duplicate reports for the same link, reason, and reporter within 24 hours are rejected",
-    responses(
-        (status = 200, description = "Report submitted successfully"),
-        (status = 400, description = "Missing required fields"),
-        (status = 404, description = "Link not found or already removed"),
-        (status = 422, description = "Link is already disabled"),
-        (status = 429, description = "Duplicate report within 24 hours"),
-    )
-)]
-pub async fn handle_report_link(mut req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    // Parse request body
-    let body: serde_json::Value = match req.json().await {
-        Ok(body) => body,
-        Err(e) => {
-            console_log!("ERROR parsing JSON: {:?}", e);
-            return Response::error(format!("Invalid JSON: {}", e), 400);
-        }
-    };
-
-    let link_id = match body.get("link_id").and_then(|v| v.as_str()) {
-        Some(id) => id.to_string(),
-        None => return Response::error("Missing 'link_id' field", 400),
-    };
-
-    let reason = match body.get("reason").and_then(|v| v.as_str()) {
-        Some(r) => r.to_string(),
-        None => return Response::error("Missing 'reason' field", 400),
-    };
-
-    let reporter_email = body.get("reporter_email").and_then(|v| v.as_str());
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-
-    // Validate that the link exists and check its status
-    let link = match db::get_active_link_by_short_code(&db, &link_id).await {
-        Ok(Some(link)) => Some(link),
-        Ok(None) => {
-            // Try by ID if short code not found
-            match db::get_link_by_id_no_auth_all(&db, &link_id).await {
-                Ok(Some(link)) => Some(link),
-                Ok(None) => {
-                    return Ok(Response::from_json(&serde_json::json!({
-                        "success": false,
-                        "message": "This link doesn't exist or has been removed.",
-                        "error_type": "link_not_found"
-                    }))?
-                    .with_status(404));
-                }
-                Err(e) => {
-                    return Response::error(format!("Database error: {}", e), 500);
-                }
-            }
-        }
-        Err(e) => {
-            return Response::error(format!("Database error: {}", e), 500);
-        }
-    };
-
-    if link.is_none() {
-        return Ok(Response::from_json(&serde_json::json!({
-            "success": false,
-            "message": "This link doesn't exist or has been removed.",
-            "error_type": "link_not_found"
-        }))?
-        .with_status(404));
-    }
-
-    let link_ref = link.unwrap();
-
-    // Check if link is already blocked or disabled
-    if matches!(link_ref.status, LinkStatus::Blocked | LinkStatus::Disabled) {
-        return Ok(Response::from_json(&serde_json::json!({
-            "success": false,
-            "message": "This link has already been disabled and cannot be reported.",
-            "error_type": "link_already_disabled"
-        }))?
-        .with_status(422));
-    }
-
-    // Get reporter info (authenticated user or email)
-    let (reporter_user_id, reporter_email_opt) = match auth::authenticate_request(&req, &ctx).await
-    {
-        Ok(user_ctx) => {
-            // Authenticated user
-            (
-                Some(user_ctx.user_id),
-                reporter_email.map(|s| s.to_string()),
-            )
-        }
-        Err(_) => {
-            // Anonymous user - use provided email
-            (None, reporter_email.map(|s| s.to_string()))
-        }
-    };
-
-    let actual_link_id = link_ref.id.clone();
-
-    // Check for duplicate reports (same link, reason, reporter within 24h)
-    let is_duplicate = db::is_duplicate_report(
-        &db,
-        &actual_link_id,
-        &reason,
-        reporter_user_id.as_deref(),
-        reporter_email_opt.as_deref(),
-    )
-    .await;
-
-    if is_duplicate.unwrap_or(false) {
-        return Response::error(
-            "You have already reported this link for the same reason within the last 24 hours",
-            429,
-        );
-    }
-
-    // Rate limiting check
-    let _client_ip = if let Ok(Some(ip)) = req.headers().get("CF-Connecting-IP") {
-        ip.to_string()
-    } else if let Ok(Some(ip)) = req.headers().get("X-Forwarded-For") {
-        ip.to_string()
-    } else {
-        "unknown".to_string()
-    };
-
-    // For now, we'll implement basic rate limiting in the application layer
-    // In production, you might want to use Redis or a dedicated rate limiting service
-
-    // Store the report in database
-    match db::create_link_report(
-        &db,
-        &actual_link_id,
-        &reason,
-        reporter_user_id.as_deref(),
-        reporter_email_opt.as_deref(),
-    )
-    .await
-    {
-        Ok(_) => {
-            // Hash sensitive fields for logging (privacy protection while maintaining correlation capability)
-            let reporter_user_id_hash = reporter_user_id.as_ref().map(|id| {
-                use hex;
-                use sha2::{Digest, Sha256}; // Add hex crate for formatting
-                let mut hasher = Sha256::new();
-                hasher.update(id.as_bytes());
-                hex::encode(hasher.finalize())
-            });
-            let reporter_email_hash = reporter_email_opt.as_ref().map(|email| {
-                use hex;
-                use sha2::{Digest, Sha256}; // Add hex crate for formatting
-                let mut hasher = Sha256::new();
-                hasher.update(email.as_bytes());
-                hex::encode(hasher.finalize())
-            });
-
-            console_log!(
-                "{}",
-                serde_json::json!({
-                    "event": "abuse_report_stored",
-                    "link_id": actual_link_id,
-                    "reason": reason,
-                    "reporter_user_id_hash": reporter_user_id_hash,
-                    "reporter_email_hash": reporter_email_hash,
-                    "level": "info"
-                })
-            );
-
-            Response::from_json(&serde_json::json!({
-                "success": true,
-                "message": "Report submitted successfully. Thank you for helping keep our platform safe."
-            }))
-        }
-        Err(e) => {
-            console_log!(
-                "{}",
-                serde_json::json!({
-                    "event": "abuse_report_failed",
-                    "error": e.to_string(),
-                    "level": "error"
-                })
-            );
-            Response::error("Failed to store report. Please try again.", 500)
-        }
-    }
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/admin/reports",
-    tag = "Admin",
-    summary = "List abuse reports",
-    params(
-        ("page" = Option<u32>, Query, description = "Page number"),
-        ("limit" = Option<u32>, Query, description = "Items per page (10-100)"),
-        ("status" = Option<String>, Query, description = "Filter by status: pending, reviewed, dismissed"),
-    ),
-    responses(
-        (status = 200, description = "Paginated list of reports"),
-        (status = 401, description = "Unauthorized"),
-        (status = 403, description = "Admin required"),
-    ),
-    security(("Bearer" = []), ("session_cookie" = []))
-)]
-pub async fn handle_admin_get_reports(req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-
-    if let Err(e) = auth::require_admin(&user_ctx) {
-        return Ok(e.into_response());
-    }
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-
-    // Parse query parameters
-    let url = req.url()?;
-    let query_pairs = url.query_pairs();
-
-    let mut page: u32 = 1;
-    let mut limit: u32 = 50;
-    let mut status_filter: Option<String> = None;
-
-    for (key, value) in query_pairs {
-        match key.as_ref() {
-            "page" => page = value.parse().unwrap_or(1),
-            "limit" => limit = value.parse().unwrap_or(50),
-            "status" => status_filter = Some(value.to_string()),
-            _ => {}
-        }
-    }
-
-    // Validate limits
-    if page < 1 {
-        page = 1;
-    }
-    limit = limit.clamp(10, 100);
-
-    match db::get_link_reports(&db, page, limit, status_filter.as_deref()).await {
-        Ok((reports, total)) => Response::from_json(&serde_json::json!({
-            "reports": reports,
-            "pagination": {
-                "page": page,
-                "limit": limit,
-                "total": total,
-                "pages": (total as f64 / limit as f64).ceil() as u32
-            }
-        })),
-        Err(e) => {
-            console_log!(
-                "{}",
-                serde_json::json!({
-                    "event": "reports_fetch_failed",
-                    "error": e.to_string(),
-                    "level": "error"
-                })
-            );
-            Response::error("Failed to retrieve reports", 500)
-        }
-    }
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/admin/reports/{id}",
-    tag = "Admin",
-    summary = "Get a single abuse report",
-    params(("id" = String, Path, description = "Report ID")),
-    responses(
-        (status = 200, description = "Report details"),
-        (status = 401, description = "Unauthorized"),
-        (status = 403, description = "Admin required"),
-        (status = 404, description = "Report not found"),
-    ),
-    security(("Bearer" = []), ("session_cookie" = []))
-)]
-pub async fn handle_admin_get_report(req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-
-    if let Err(e) = auth::require_admin(&user_ctx) {
-        return Ok(e.into_response());
-    }
-
-    let report_id = ctx
-        .param("id")
-        .ok_or_else(|| Error::RustError("Missing report ID".to_string()))?
-        .to_string();
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-
-    match db::get_link_report_by_id(&db, &report_id).await {
-        Ok(report) => Response::from_json(&report),
-        Err(_) => Response::error("Report not found", 404),
-    }
-}
-
-#[utoipa::path(
-    put,
-    path = "/api/admin/reports/{id}",
-    tag = "Admin",
-    summary = "Update report status",
-    params(("id" = String, Path, description = "Report ID")),
-    responses(
-        (status = 200, description = "Report updated"),
-        (status = 400, description = "Invalid status"),
-        (status = 401, description = "Unauthorized"),
-        (status = 403, description = "Admin required"),
-    ),
-    security(("Bearer" = []), ("session_cookie" = []))
-)]
-pub async fn handle_admin_update_report(
-    mut req: Request,
-    ctx: RouteContext<()>,
-) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-
-    if let Err(e) = auth::require_admin(&user_ctx) {
-        return Ok(e.into_response());
-    }
-
-    let report_id = ctx
-        .param("id")
-        .ok_or_else(|| Error::RustError("Missing report ID".to_string()))?
-        .to_string();
-
-    // Parse request body
-    let body: serde_json::Value = match req.json().await {
-        Ok(body) => body,
-        Err(e) => return Response::error(format!("Invalid JSON: {}", e), 400),
-    };
-
-    let status = match body.get("status").and_then(|v| v.as_str()) {
-        Some(s) if s == "reviewed" || s == "dismissed" => s.to_string(),
-        Some(_) => {
-            return Response::error("Invalid status. Must be 'reviewed' or 'dismissed'", 400);
-        }
-        None => return Response::error("Missing 'status' field", 400),
-    };
-
-    let admin_notes = body.get("admin_notes").and_then(|v| v.as_str());
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-
-    match db::update_link_report_status(&db, &report_id, &status, &user_ctx.user_id, admin_notes)
-        .await
-    {
-        Ok(_) => Response::from_json(&serde_json::json!({
-            "success": true,
-            "message": "Report status updated successfully"
-        })),
-        Err(e) => {
-            console_log!(
-                "{}",
-                serde_json::json!({
-                    "event": "report_update_failed",
-                    "error": e.to_string(),
-                    "level": "error"
-                })
-            );
-            Response::error("Failed to update report status", 500)
-        }
-    }
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/admin/reports/pending/count",
-    tag = "Admin",
-    summary = "Get pending reports count",
-    responses(
-        (status = 200, description = "Count of pending reports"),
-        (status = 401, description = "Unauthorized"),
-        (status = 403, description = "Admin required"),
-    ),
-    security(("Bearer" = []), ("session_cookie" = []))
-)]
-pub async fn handle_admin_get_pending_reports_count(
-    req: Request,
-    ctx: RouteContext<()>,
-) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-
-    if let Err(e) = auth::require_admin(&user_ctx) {
-        return Ok(e.into_response());
-    }
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-
-    match db::get_pending_reports_count(&db).await {
-        Ok(count) => Response::from_json(&serde_json::json!({
-            "count": count
-        })),
-        Err(e) => {
-            console_log!(
-                "{}",
-                serde_json::json!({
-                    "event": "reports_count_failed",
-                    "error": e.to_string(),
-                    "level": "error"
-                })
-            );
-            Response::error("Failed to get pending reports count", 500)
-        }
-    }
-}
-
-#[utoipa::path(
     get,
     path = "/api/admin/billing-accounts",
     tag = "Admin",
